requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-bcb>=0.2.0
plotly>=5.15.0
jupyterlab>=4.0.0